refactor: share frontend money input helpers
